pyproject.toml
[project]
name = "fastapi-authentication"
version = "0.1.0"
description = ""
authors = [
    {name = "Manish Dhull",email = "manishdhull1990@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "ldap3 (>=2.9.1,<3.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "starlette (==0.40.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pymysql (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "redis[async] (>=6.2.0,<7.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)"
]

[tool.poetry]
packages = [{include = "fastapi_authentication", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
